web/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
apispec==6.6.1
astroid==3.1.0
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
flask-accept==0.0.6
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.3
limits==3.11.0
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mccabe==0.7.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.0
platformdirs==4.2.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.1.0
pylint-flask==0.6
pylint-plugin-utils==0.8.2
pymongo==4.7.0
python-dotenv==1.0.1
python-json-logger==2.0.7
rich==13.7.1
SQLAlchemy==2.0.29
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
webargs==8.4.0
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
click